Take DashboardView out of internal
* move the files up the structure
* review all the release tags and links

JIRA: RAIL-2956
