+++ libs/sdk-ui-ext/src/dashboardView/defaultDashboardThemeModifier.ts
@@ -2,7 +2,12 @@
 import cloneDeep from "lodash/cloneDeep";
 import { ITheme } from "@gooddata/sdk-backend-spi";
 
-export const defaultThemeModifier = (theme: ITheme): ITheme => {
+/**
+ * Default modifier applied to any theme passed to DashboardView
+ * @param theme - theme to modify
+ * @beta
+ */
+export const defaultDashboardThemeModifier = (theme: ITheme): ITheme => {
     if (theme?.dashboards?.content?.kpiWidget?.kpi) {
         // duplicate dashboard specific kpi customization to the generic kpi key
         const themeWithKpi = cloneDeep(theme);
