+++ libs/sdk-ui-ext/src/dashboardView/index.ts
@@ -1,0 +1,9 @@
+// (C) 2020-2021 GoodData Corporation
+import { DashboardViewErrorBoundary, IDashboardViewErrorBoundaryState } from "./DashboardViewErrorBoundary";
+export { DashboardViewErrorBoundary as DashboardView, IDashboardViewErrorBoundaryState };
+
+export { defaultDashboardThemeModifier } from "./defaultDashboardThemeModifier";
+export { mergeFiltersWithDashboard } from "./mergeFiltersWithDashboard";
+
+export * from "./types";
+export * from "./hooks";
