+++ libs/sdk-ui-ext/src/index.ts
@@ -1,4 +1,5 @@
 // (C) 2007-2021 GoodData Corporation
 
 export { clearInsightViewCaches, clearDashboardViewCaches } from "./dataLoaders";
+export * from "./dashboardView";
 export * from "./insightView";
