+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/KpiView/alertManipulationHooks/useAlertDeleteHandler.ts
@@ -1,7 +1,7 @@
 // (C) 2021 GoodData Corporation
 import { useState } from "react";
 import { IWidgetAlert } from "@gooddata/sdk-backend-spi";
-import { useDeleteWidgetAlert } from "../../../hooks/internal";
+import { useDeleteWidgetAlert } from "../../../hooks";
 import { KpiAlertOperationStatus } from "../../../types";
 import { useAlerts } from "../../contexts";
 import { IUseAlertManipulationHandlerConfig } from "./types";
