--- libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/index.ts
@@ -1,6 +1,0 @@
-// (C) 2020-2021 GoodData Corporation
-import { DashboardViewErrorBoundary } from "./DashboardViewErrorBoundary";
-export { DashboardViewErrorBoundary as DashboardView };
-export { IDashboardViewProps, DashboardLayoutTransform } from "./types";
-export { KpiView } from "./KpiView";
-export { defaultThemeModifier } from "./defaultThemeModifier";
