+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/index.ts
@@ -1,5 +1,16 @@
 // (C) 2021 GoodData Corporation
-export * from "./useDashboard";
-export * from "./useDashboardAlerts";
-export * from "./useDashboardPdfExporter";
-export * from "./useDashboardWidgetExecution";
+export * from "./useAttributesWithDrillDown";
+export * from "./useColorPalette";
+export * from "./useCurrentUser";
+export * from "./useDashboardInsightExecution";
+export * from "./useDashboardKpiExecution";
+export * from "./useDashboardLayoutData";
+export * from "./useDeleteWidgetAlert";
+export * from "./useExportDashboardToPdf";
+export * from "./useKpiData";
+export * from "./useSaveOrUpdateWidgetAlert";
+export * from "./useSaveScheduledMail";
+export * from "./useUpdateWidgetAlert";
+export * from "./useUserWorkspacePermissions";
+export * from "./useUserWorkspaceSettings";
+export * from "./useWorkspaceUsers";
