+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/useAttributesWithDrillDown.ts
@@ -9,8 +9,8 @@ import {
     UseCancelablePromiseState,
     useWorkspace,
 } from "@gooddata/sdk-ui";
-import { attributesWithDrillDownDataLoaderFactory } from "../../../../dataLoaders";
-import { backendInvariant, workspaceInvariant } from "../utils";
+import { attributesWithDrillDownDataLoaderFactory } from "../../../dataLoaders";
+import { backendInvariant, workspaceInvariant } from "../../../utils";
 
 interface IAttributesWithDrillDownConfig
     extends UseCancelablePromiseCallbacks<
