+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/types.ts
@@ -1,11 +1,5 @@
 // (C) 2007-2021 GoodData Corporation
 import { IMeasureDescriptor } from "@gooddata/sdk-backend-spi";
-import {
-    IAbsoluteDateFilter,
-    IRelativeDateFilter,
-    IPositiveAttributeFilter,
-    INegativeAttributeFilter,
-} from "@gooddata/sdk-model";
 
 export interface IKpiResult {
     measureFormat: string;
@@ -20,13 +14,3 @@ export interface IKpiAlertResult {
 }
 
 export type KpiAlertOperationStatus = "idle" | "inProgress" | "error";
-
-/**
- * Supported dashboard filter type.
- * @alpha
- */
-export type IDashboardFilter =
-    | IAbsoluteDateFilter
-    | IRelativeDateFilter
-    | IPositiveAttributeFilter
-    | INegativeAttributeFilter;
